[TSK-276] refactor: consolidate terms of use and privacy policy routes into init file
